+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatButtonUI.java
@@ -573,7 +573,10 @@ public class FlatButtonUI
 	public static Color buttonStateColor( Component c, Color enabledColor, Color disabledColor,
 		Color focusedColor, Color hoverColor, Color pressedColor )
 	{
-		if(c != null && !c.isEnabled() )
+		if( c == null )
+			return enabledColor;
+
+		if( !c.isEnabled() )
 			return disabledColor;
 
 		if( c instanceof AbstractButton ) {
@@ -586,7 +589,7 @@ public class FlatButtonUI
 				return hoverColor;
 		}
 
-		if( c != null && focusedColor != null && isFocusPainted( c ) && FlatUIUtils.isPermanentFocusOwner( c ) )
+		if( focusedColor != null && isFocusPainted( c ) && FlatUIUtils.isPermanentFocusOwner( c ) )
 			return focusedColor;
 
 		return enabledColor;